Right pane (item-pane) can be collapsed
This belongs to issue #509 and was asked several times in the forum.

 * Add attribute `collapse="after"` to the splitter and add a grippy element.
 * To remember its state after restart I copied `zotero-persist="state"` which seems to do the job.
 * Update the function `updateToolbarPosition` in zotero/zoteroPane.js:
   * The width of the items-toolbar is corrected if the left pane is collapsed (and all the icons are grouped on the left margin together).
   * If the right pane is collapsed, then the items-toolbar is made flexible (as the item-toolbar is already). As a result the icons from the two toolbars are spreaded evenly over the place.

I tested it in Firefox. Moreover, I installed the zotero-item-history plugin on top of it (this gives new icons to the toolbars). Everything works fine for me. One thing I encountered: if the width of the collections-pane is maximized and the right pane is collapsed, then the icons will fall out on the right. However, I hope that this will not occur often (if ever) in practice.